+++ include/editor.h
@@ -21,11 +21,11 @@ typedef struct {
 extern Editor editor;
 
 void Editor_Initialize();
-void Editor_Update(double dt);
+void Editor_Update(float dt);
 void Editor_Select_Edit_Point(VectorInt mousePos, bool isEnemyRoute);
 void Editor_Select_Enemy(VectorInt mousePos);
-void Editor_Place(double x, double y, bool isEnemyRoute);
-void Editor_Drag(double x, double y);
+void Editor_Place(float x, float y, bool isEnemyRoute);
+void Editor_Drag(float x, float y);
 void Editor_Drag_Enemy(Vector mouseGamePos, VectorInt mousePos);
 void Editor_Delete_Point(bool isEnemyRoute);
 void Editor_Delete_Enemy();
